sync_zeit: store screenshot

+++ sync_zeit.sh
@@ -2,7 +2,10 @@
 set -exu
 mkdir -p work && cd work || exit 1
 
-fetch_zeit.py --formats="epub pdf"
+if ! fetch_zeit.py --formats="epub pdf"; then
+    mv zeit_login_error.png /tmp
+    exit 1
+fi
 rclone copy "$PWD" zeitung:
 rm ./*.epub
 rmapi mput Zeitung
